test/fedbird/requirements2.txt
numpy
keras
tensorflow
sklearn
pandas
pillow
matplotlib

